Implements the current playlist feature
If a playlist is loaded, it's id is stored on the user row. When the
page loads, the playlist that will be loaded is the latest playlist that
the user played. When a new user is created, a playlist (named default)
is also created and this playlist is set as the user's current playlist.

When a playlist (different than default) is loaded and the user adds new
tracks, the tracks are added to the playlist loaded, not the default.